LPD-40070 Use a non-recursive approach by default
In the first version we were extracting inline handlers from all nodes in
the HTML but that may introduce issues that CSP is trying to prevent.

So now by default we will be only extracting handlers from the top level
nodes of the HTML unless recursive=true is passed to the JSP tag or Util
class.

This will also enhance performance.